# docs/openapi.yaml
openapi: 3.0.3
info:
  title: نظام إدارة محل السبائك والمشغولات — API
  description: |
    REST API لنظام إدارة محل الذهب والفضة: مخزون مركزي، تسعير يومي، فروع متعددة،
    نقاط بيع (POS)، حجوزات، جرد، شيفتات، تقارير، ومزامنة مع WooCommerce.

    ## المصادقة
    - جميع النقاط تحتاج `Authorization: Bearer <token>` ما عدا `/api/auth/login` و `/api/health`.
    - توكن JWT (خوارزمية HS256، صلاحية 12 ساعة) يُحصل عليه من `/api/auth/login`.
    - لكل مسؤول صلاحيات (permissions) تُحمّل مع كائن الموظف عند الدخول.
      النقاط المحمية بصلاحية محددة موثّقة بـ «Requires permission» في الوصف،
      وترجع `403 {error:"forbidden"}` عند غيابها.

    ## الأخطاء
    جميع الأخطاء تُرجع JSON على الصورة: `{"error": "<code>"}` مع اختياري
    `message` (نص عربي قابل للعرض) و `detail` (لأخطاء الاتصال) و `refs` (للحذف المقيّد).
    الكود الداخلي في `error` ثابت وقابل للبرمجة عليه.

    ## توكن الدخول
    استخدم `/api/auth/login` بالحسابات الأولية (PIN: `1234`):
    | username | الدور |
    |---|---|
    | `manager` | مدير المحل |
    | `cashier` | كاشير |
    | `social` | سوشيال |
  version: 1.0.0
  contact:
    name: Jewelry Store System
servers:
  - url: http://localhost:4001
    description: تطوير (Vite proxy)
  - url: https://your-domain.example
    description: إنتاج
...
security:
  - bearerAuth: []
paths:
...
  /api/stock-limits:
    get:
      tags: [StockLimits]
      summary: حدود المخزون (min/max) لكل فرع+معدن+عيار
      responses:
        '200':
          description: الحدود
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StockLimit' }
    post:
      tags: [StockLimits]
      summary: إنشاء/تحديث حد مخزون (upsert بالفرع+معدن+عيار)
      description: "Requires permission: `stockcount.manage`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locationId, metalType]
              properties:
                locationId: { type: integer }
                metalType: { type: string, enum: [gold, silver] }
                carat: { type: string }
                minQty: { type: integer, default: 0 }
                maxQty: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockLimit'
  /api/stock-limits/{id}:
    delete:
      tags: [StockLimits]
      summary: حذف حد مخزون
      description: "Requires permission: `stockcount.manage`"
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          $ref: '#/components/responses/Ok'

  # ------------------------------------------------------------------
  # Invoices
  # ------------------------------------------------------------------
...
  /api/reservations:
    get:
      tags: [Reservations]
      summary: قائمة الحجوزات (آخر 300)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, completed, cancelled] } }
      responses:
        '200':
          description: الحجوزات
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reservation' }
    post:
      tags: [Reservations]
      summary: إنشاء حجز بالعربون
      description: |
        Requires permission: `reservation.manage`.
        يثبّت `reserved_qty` للقطعة. 409 `items.not_available` إن كانت الكمية غير كافية.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/reservations/{id}/cancel:
    post:
      tags: [Reservations]
      summary: إلغاء حجز (يُطلق الكمية المحجوزة)
      description: "Requires permission: `reservation.manage`. 409 `reservation.not_active` إن لم يكن الحجز نشطًا."
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '409':
          $ref: '#/components/responses/Conflict'

  # ------------------------------------------------------------------
  # Shifts
  # ------------------------------------------------------------------
...
  /api/customers/{id}/reservations:
    get:
      tags: [Customers]
      summary: حجوزات عميل (آخر 100)
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reservation' }

  # ------------------------------------------------------------------
  # Payment methods
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: توكن JWT من /api/auth/login (صلاحية 12 ساعة)
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
      description: معرّف رقمي
  responses:
    Ok:
      description: نجاح
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean, example: true }
...
    NotFound:
      description: غير موجود
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: تعارض (حالة غير صالحة أو مرجع محجوب)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: طلب غير صالح
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: كود خطأ ثابت قابل للبرمجة، مثال notfound / items.in_use }
        message: { type: string, description: نص عربي قابل للعرض (اختياري) }
        detail: { type: string, description: تفاصيل إضافية (اختياري) }
        refs:
          type: object
          description: إحصاءات المراجع للحذف المقيّد (اختياري)

    # ---- Auth ----
...
    StockLimit:
      type: object
      properties:
        id: { type: integer }
        locationId: { type: integer }
        metalType: { type: string, enum: [gold, silver] }
        carat: { type: string, nullable: true }
        minQty: { type: integer }
        maxQty: { type: integer, nullable: true }

    # ---- Invoices ----
...
    Reservation:
      type: object
      properties:
        id: { type: integer }
        itemId: { type: integer }
        itemCode: { type: string }
        itemName: { type: string }
        quantity: { type: integer }
        customerId: { type: integer, nullable: true }
        customerName: { type: string }
        customerPhone: { type: string, nullable: true }
        downPayment: { type: number }
        totalValue: { type: number }
        remainingDue: { type: number }
        reservedBy: { type: integer, nullable: true }
        reservedByName: { type: string, nullable: true }
        reservedAt: { type: string, format: date-time }
        status: { type: string, enum: [active, completed, cancelled] }
        invoiceId: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
    ReservationCreate:
      type: object
      required: [itemId, downPayment, totalValue]
      properties:
        itemId: { type: integer }
        quantity: { type: integer, minimum: 1, default: 1 }
        customerId: { type: integer }
        customerName: { type: string, description: يُستخدم إن لم يُمرَّر customerId }
        customerPhone: { type: string }
        downPayment: { type: number, description: العربون }
        totalValue: { type: number }
        notes: { type: string }

    # ---- Shifts ----
...